Praca domowa - moduł 5

+++ WebAPI/Controllers/V1/PostController.cs
@@ -1,21 +1,24 @@
 ﻿using Application.DTO;
 using Application.Interfaces;
+using Infrastructure.Identity;
 using Microsoft.AspNet.OData;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using WebAPI.Filters;
 using WebAPI.Helpers;
 using WebAPI.Wrappers;
 
 namespace WebAPI.Controllers.V1
-{
-    [ApiExplorerSettings(IgnoreApi = true)]
+{   
     [Route("api/[controller]")]
     [ApiVersion("1.0")]
+    [Authorize]
     [ApiController]
     public class PostController : Controller
     {
@@ -33,6 +36,7 @@ namespace WebAPI.Controllers.V1
         }
 
         [SwaggerOperation(Summary ="Retrieves paged posts")]
+        [AllowAnonymous]
         [HttpGet]
         public async Task<IActionResult> Get([FromQuery] PaginationFilter paginationFilter, [FromQuery] SortingFilter sortingFilter,
                                               [FromQuery] string filteredBy="")
@@ -50,6 +54,7 @@ namespace WebAPI.Controllers.V1
         }
 
         [SwaggerOperation(Summary ="Retrieves all posts")]
+        [Authorize(Roles = UserRoles.SuperUserOrAdmin)]        
         [EnableQuery]
         [HttpGet("[action]")]
         public IQueryable<PostDto> GetAll()
@@ -58,6 +63,7 @@ namespace WebAPI.Controllers.V1
         }
 
         [SwaggerOperation(Summary = "Retrieves post")]
+        [AllowAnonymous]
         [HttpGet("{id}")]
         public async Task<IActionResult> Get(int id)
         {
@@ -81,10 +87,11 @@ namespace WebAPI.Controllers.V1
         }
 
         [SwaggerOperation(Summary = "Create new post")]
+        [Authorize(Roles = UserRoles.SuperUserOrUser)]
         [HttpPost]
         public async Task<IActionResult> Create(CreatePostDto newPost)
         {
-            var post = await _postService.AddNewPostAsync(newPost);
+            var post = await _postService.AddNewPostAsync(newPost, User.FindFirstValue(ClaimTypes.NameIdentifier));
             return Created($"api/posts/{post.Id}", new Response<PostDto>(post));
         }
 
@@ -92,14 +99,30 @@ namespace WebAPI.Controllers.V1
         [HttpPut]
         public async Task<IActionResult> Update(UpdatePostDto updatePost)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(updatePost.Id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+            if(!userOwnsPost)
+            {
+                return BadRequest(new Response<bool>() { Succeeded = false, Message = "You do not own this post" });
+            }
             await _postService.UpdatePostAsync(updatePost);
             return NoContent();
         }
 
         [SwaggerOperation(Summary = "Delete post")]
+        [Authorize(Roles = UserRoles.SuperUserOrAdminOrUser)]
         [HttpDelete]
         public async Task<IActionResult> Delete(int id)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+            var isAdmin = User.IsInRole(UserRoles.Admin);
+            var isSuperUser = User.IsInRole(UserRoles.SuperUser);
+
+            if (!isSuperUser && !isAdmin && !userOwnsPost)
+            {
+                return BadRequest(new Response<bool>() { Succeeded = false, Message = "You do not own this post" });
+            }
+
             await _postService.DeletePostAsync(id);
             return NoContent();
         }
